Create Step 1 - Process the Data
In this step we take the master file and process the data, creating logarithms for the Distance to the nearest Highway Interchange and dummy variables for Building Class (A, B or C), Sale Status (Sold/Under Contract), Location Type (CBD/Urban/Suburban), Number of Floors (1-8 and 13) and quarter (2019Q1 through 2021Q2). Target columns for analysis are defined as: Price per SF, Distance to the Nearest Interchange and Ceiling Height (in decimals), plus the dummy variables. A new csv file with the processed data is created and saved.