Create Step 3 - Interval Value Model
This third step estimates the value of the subject property in dollars and dollars per square foot and also gives a value range in dollars that is 80% reliable using the .pkl file produced in Step 2. A dictionary comprehension was tried for the dummy variables for quarter (2019Q1 - 2021Q2), MSA (Numerical code), Sale Status (Sold vs Under Contract) , Building Class (A, B or C), Location Type (Urban, Suburban or CBD) and Number of Floors (1-8 and 13) but the model produced an error or gave a null result. A positive and reasonable result was achieved by creating a new dictionary for these dummy variables where the analyst must enter 1 for the attibutes they want to select and 0 for the attribiutes they want to deselect. Building area in square feet must be entered at lines 13, 99, 100, 101 and 109, as using BuildingSF as a variable after entering in in Line 13 didn't seem to work. The model produces a low, high and mean value on both a dollars and dollars per square foot basis and then expresses the value as a range.