Source Data
The source data is a combination of an export from a common real estate database used by appraisers, brokers and other real estate professionals. I requested data according to the following parameters: 

Property Type: Industrial
Size: 50,000 SF+
Location: United States, East of the Mississippi River
Time Frame: January 2019 - present
File type: Excel

I received 39 columns and 7,083 rows of data, each representing sales either closed or in contract as of April 15, 2021. 

I also utilized data from the US Department of Labor, Office of Worker's Compensation Programs Geographic Practice Cost Index (GCPI). The Practice Expense GCPI adjusts for geographic variation in the direct costs of providing services and the indirect costs of maintaining a clinical practice, including administrative and clinical staff compensation (salary and benefits), rent, and supplies and equipment. It accounts for 43.7% of geographic adjustment on average. The GCPI is calculated for every zip code in the United States and more granular cost index data could not be found within the allotted time frame, so the Practice Expense GCPI was used as a proxy for a Cost of Living Index. In the next version, I would like to more explicitly link this cost index to zip code within the model instead of it being a separate variable. 

For the geodata, I found a video on YouTube called ": Calculating Distance In Excel" (https://www.youtube.com/watch?v=cpUVV8q7WNo). There was an accompanying Excel file containing the following formula to convert Zip Codes to geodata and then compute the distance between two points: 
	(=ACOS(COS(RADIANS(90-B2))*COS(RADIANS(90-H2))+SIN(RADIANS(90-B2))*SIN(RADIANS(90-H2))*COS(RADIANS(C2-I2)))*3958.8)

Once I did that I found a list of major highway interchanges in the Eastern United States, found the zip codes they were located in, and then also entered them into the Excel model to calculate the distance between each subject and each highway interchange. At the end of the line I used an Excel formula to find the minimum distance between the subject and the nearest interchange, then used a series of IF statements to create a formula for the model to identify in which zip code the nearest highway interchange is located. Lastly, a V-lookup was used to pair the subject zip codes with the correct MSA (Metropolitan Statistical Area) and GCPI. 

Many of the columns in the original exported data were irrelevant, so many of the columns were deleted. The remaining columns were: Sale Price, Price Per SF, Sale Date, Sale Status, Secondary Type, Land Area AC, Number Of Floors, Ceiling Height, Location Type, Property Zip Code

Meanwhile, the following new columns were created: MSA, Nearest Major Highway Interchange, Distance to Nearest Interchange

Ceiling height was in traditional feet and inches. This had to be converted to decimals using a formula from Microsoft Answers. Any ranges were averaged by using code to split the data in the columns and then creating a third column to average the first and second numbers after which all the ceiling height data was then consolidated into one column.  Rows that did not include Ceiling Height data were deleted, as well as multiple property sales, bringing the final number of rows to 3,515, each representing a sale of a single property. 

At this point, the Excel worksheet was converted to .csv and the data was entered into Model Step 0 to process and describe the data before analysis.